+++ src/components/MessageBar/MessageBar.tsx
@@ -10,9 +10,6 @@ import { Button as MessageBarButton } from '../Button/Button';
 import { getClassNames } from './MessageBar.classNames';
 import { MessageBarProps, MessageBarState } from './MessageBar.types';
 
-/*
- * visibleName MessageBar (Varsler)
- */
 export class MessageBar extends React.PureComponent<
   MessageBarProps,
   MessageBarState