+++ src/components/TopBanner/TopBanner.spec.js
@@ -41,9 +41,10 @@ describe('TopBanner komponent', () => {
     expect(wrapper.find('h1').html()).toContain('Ekstern publikumsløsning');
     expect(wrapper.html()).toContain('contentWrapper');
     expect(wrapper.html()).toContain('headerMain');
-    expect(wrapper.find('.ms-Icon').prop('data-icon-name')).toEqual('Back');
+    expect(wrapper.find('.ms-Icon').prop('data-icon-name')).toEqual(
+      'ArrowBack'
+    );
     expect(wrapper.find('ImageBase').prop('height')).toEqual(68);
-    expect(wrapper.find('img').prop('alt')).toEqual('Skatteetaten logo');
   });
 
   it('rendrer TopBanner med kompakt visning for eksterne arbeidsflater ', () => {