+++ src/components/utils/loadTheme.ts
@@ -8,7 +8,7 @@ import {
   Palette as palette,
   Fonts as fonts,
   SkeIcons,
-  MdIcons,
+  AliasIcons,
   PaletteProps,
 } from '../utils/index';
 // @ts-ignore TODO
@@ -21,4 +21,4 @@ loadTheme(theme);
 // @ts-ignore TODO
 registerIcons(SkeIcons);
 // @ts-ignore TODO
-registerIcons(MdIcons);
+registerIcons(AliasIcons);