+++ CHANGELOG.md
@@ -15,6 +15,15 @@
 - Eksport til ESM: Endrer fra umd til esm eksport (//TO-DO reaktivere import-maps)
 - Reorganisering av import/export til å fungere med moduler (//TO-DO alle doc-sidene må retestes) //TO-DO fjerne entry.ts i /lib etter compile
 
+### **v.4.1.5 - 01.10.2021**
+
+- SearchField: fixed keyboard navigation when using a filtered list.
+
+### **v.4.1.4 - 22.09.2021**
+
+- FileUploader: added more supported file types. Fixes an isse with the isLoading-prop that resulted in a console error.
+- SearchField: fixes type specification on props searchShortcutKeys and searchIconTitle ('string' to string).
+
 ### **v.4.1.3 - 06.09.2021**
 
 - Revert change from v.4.1.0, compiling to es modules, due to problems with build pipelines.